+++ tests/templates/test_classes_table.py
@@ -68,7 +68,7 @@ def rows(table: etree.Element) -> List[List[str]]:
     rows = []
     for row in table.findall(path=".//tbody//tr"):
         content = []
-        for cell in row.findall(path=".//td"):
+        for cell in row.xpath(_path=".//th | .//td"):
             content.append(cell.text)
         rows.append(content)
     return rows
@@ -179,6 +179,9 @@ def test_classes_table_with_several_classes(render, event: EventType):
         "Penalty time limit",
     ]
     assert rows(table) == [
+        [
+            "Classes\xa0\xa0(3)",
+        ],
         [
             "Elite",
             "Elite",