+++ tests/templates/test_clubs_table.py
@@ -46,7 +46,7 @@ def rows(table: etree.Element) -> List[List[str]]:
     rows = []
     for row in table.findall(path=".//tbody//tr"):
         content = []
-        for cell in row.findall(path=".//td"):
+        for cell in row.xpath(_path=".//th | .//td"):
             content.append(cell.text)
         rows.append(content)
     return rows
@@ -83,6 +83,9 @@ def test_clubs_table_with_several_clubs(render):
         "Name",
     ]
     assert rows(table) == [
+        [
+            "Clubs\xa0\xa0(2)",
+        ],
         [
             "OL Bundestag",
         ],