+++ tests/templates/test_courses_table.py
@@ -65,7 +65,7 @@ def rows(table: etree.Element) -> List[List[str]]:
     rows = []
     for row in table.findall(path=".//tbody//tr"):
         content = []
-        for cell in row.findall(path=".//td"):
+        for cell in row.xpath(_path=".//th | .//td"):
             content.append(cell.text)
         rows.append(content)
     return rows
@@ -120,6 +120,9 @@ def test_courses_table_with_several_courses(render, event: EventType):
         "Controls",
     ]
     assert rows(table) == [
+        [
+            "Courses\xa0\xa0(2)",
+        ],
         [
             "Bahn B",
             None,